Convert the TraceStateBuilder into an interface.
